layout and folder for root created
