+++ app/(main)/page.tsx
@@ -1,0 +1,13 @@
+import Image from 'next/image'
+
+export default function Home() {
+  return (
+    <>
+      <main className="flex flex-col items-center justify-between p-24">
+        <div className="mb-32 grid text-center lg:max-w-5xl lg:w-full lg:mb-0 lg:grid-cols-4 lg:text-left">
+          this is the test content for the website and more is comming soon.
+        </div>
+      </main>
+    </>
+  )
+}
